+++ tower-oauth2-resource-server/tests/common/jwks.rs
@@ -1,4 +1,8 @@
 use serde::Serialize;
+use wiremock::{
+    matchers::{method, path},
+    Mock, MockServer, ResponseTemplate,
+};
 
 use crate::common::rsa::RsaKey;
 
@@ -30,4 +34,13 @@ pub fn build_jwks(keys: &[(&str, &RsaKey)]) -> Jwks {
         })
         .collect::<Vec<_>>();
     Jwks { keys }
+}
+
+pub async fn mock_jwks(mock_server: &MockServer, issuer_path: &str, keys: &[(&str, &RsaKey)]) {
+    let jwks = build_jwks(keys);
+    Mock::given(method("GET"))
+        .and(path(format!("{}/jwks", issuer_path)))
+        .respond_with(ResponseTemplate::new(200).set_body_json(jwks))
+        .mount(mock_server)
+        .await;
 }
