+++ tower-oauth2-resource-server/tests/common/oidc.rs
@@ -1,7 +1,25 @@
 use serde::Serialize;
+use wiremock::{
+    matchers::{method, path},
+    Mock, MockServer, ResponseTemplate,
+};
 
 #[derive(Serialize)]
 pub struct OpenIdConfig {
     pub issuer: String,
     pub jwks_uri: String,
+}
+
+pub async fn mock_oidc(mock_server: &MockServer, issuer_path: &str) {
+    Mock::given(method("GET"))
+        .and(path(format!(
+            "{}/.well-known/openid-configuration",
+            issuer_path
+        )))
+        .respond_with(ResponseTemplate::new(200).set_body_json(OpenIdConfig {
+            issuer: format!("{}{}", &mock_server.uri(), issuer_path),
+            jwks_uri: format!("{}{}/jwks", &mock_server.uri(), issuer_path),
+        }))
+        .mount(mock_server)
+        .await;
 }
